Add Remote Controller controll function
In Remote Controller mode You can control the Snake by pressing the "ARROW" buttons on the Remote Controller.
The original plan was that You can switch between the modes by pressing the JoySitck or pressing the "OK" button. But the JoyStick's button didn't work well, so this is the reason why You can see the "the JoyStick's button dosn't work well so switch back with ok button" or similar comments on the code.
The program starts with the JoyStick mode.